Move doc() and man() into a new docs module
